Move controller into controller.h
